Fix wrong revision ID when approving while browsing old revisions
When browsing through a page revisions, "Previous" and "Next revision" links
are of the form "?title=PageName&direction=prev/next&oldid=…",
where "oldid" corresponds to the ID of the revision where these links appear
(not the target revision).
When actually displaying the target revision, "oldid" thus does not correspond
to the current revision.
However, setArticleHeader function considers that $request->getInt( 'oldid' )
returns the ID of the current revision when building the "Approve this
revision" link.
The wrong revision is thus approved.
We fix this bug by building the "Approve this revision" link in other functions
were the current revision ID is provided.